# specs/2-api-server-docker/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Quantum Circuit Task Queue API
  description: |
    REST API for submitting quantum circuit execution tasks and retrieving results.

    This is a stubbed implementation providing foundational endpoints for parallel development.
    Actual task execution and persistence will be added in subsequent features.

    **Key Features:**
    - Submit quantum circuits for async execution
    - Query task status and results
    - Health check for monitoring

    **Constitution Compliance:**
    - Principle 3 (Observable): Correlation IDs for request tracing
    - Principle 7 (Separation of Concerns): API layer does not execute tasks
    - Principle 8 (Production Patterns): RESTful design, standard HTTP codes
  version: 1.0.0
  contact:
    name: Quantum Task Queue Team
  license:
    name: MIT

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: http://api.example.com
    description: Production server (example)

tags:
  - name: tasks
    description: Task submission and status operations
  - name: health
    description: Service health and monitoring

paths:
  /tasks:
    post:
      tags:
        - tasks
      summary: Submit quantum circuit task
      description: |
        Submit a quantum circuit for asynchronous execution. Returns a unique task ID
        that can be used to query execution status and results.

        **Stub Behavior:** Task is not actually executed or persisted. A UUID is generated
        and returned for testing client integration.
      operationId: submitTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskSubmitRequest'
            examples:
              bell_state:
                summary: Bell state circuit
                value:
                  circuit: "OPENQASM 3; qubit[2] q; h q[0]; cnot q[0], q[1]; measure q;"
              hadamard:
                summary: Single qubit Hadamard
                value:
                  circuit: "OPENQASM 3; qubit q; h q; measure q;"
      responses:
        '200':
          description: Task submitted successfully
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSubmitResponse'
              examples:
                success:
                  value:
                    task_id: "550e8400-e29b-41d4-a716-446655440000"
                    message: "Task submitted successfully."
                    correlation_id: "abc123-def456-789012"
        '400':
          description: Invalid request (validation error)
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_circuit:
                  value:
                    error: "Validation failed"
                    details:
                      circuit: "Field required"
                    correlation_id: "abc123-def456"
                empty_circuit:
                  value:
                    error: "Validation failed"
                    details:
                      circuit: "String should have at least 1 character"
                    correlation_id: "abc123-def456"
        '415':
          description: Unsupported Media Type (non-JSON request)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                server_error:
                  value:
                    error: "Internal server error"
                    correlation_id: "abc123-def456"

  /tasks/{task_id}:
    get:
      tags:
        - tasks
      summary: Get task status
      description: |
        Retrieve the current status and results of a previously submitted task.

        **Stub Behavior:** All valid UUIDs return status "pending" with a fixed message.
        Actual status lookup will be implemented when database integration is added.
      operationId: getTaskStatus
      parameters:
        - name: task_id
          in: path
          required: true
          description: Unique task identifier (UUID v4)
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Task status retrieved successfully
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatusResponse'
              examples:
                pending:
                  summary: Task pending
                  value:
                    status: "pending"
                    message: "Task is still in progress."
                    correlation_id: "xyz789-uvw456"
                completed:
                  summary: Task completed (future)
                  value:
                    status: "completed"
                    result:
                      "0": 512
                      "1": 512
                    correlation_id: "xyz789-uvw456"
        '400':
          description: Invalid task ID format
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_uuid:
                  value:
                    error: "Invalid task ID format. Expected UUID v4."
                    correlation_id: "xyz789-uvw456"
        '404':
          description: Task not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  value:
                    error: "Task not found."
                    correlation_id: "xyz789-uvw456"
        '500':
          description: Internal server error
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      tags:
        - health
      summary: Health check
      description: |
        Check service health and readiness. Used by monitoring systems and container
        orchestration for liveness/readiness probes.

        **Response Time:** < 100ms (typically < 10ms)

        **Stub Behavior:** Always returns "healthy" while server is running. Future
        versions will check database and queue connectivity.
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckResponse'
              examples:
                healthy:
                  value:
                    status: "healthy"
                    timestamp: "2025-12-28T12:00:00Z"

components:
  schemas:
    TaskSubmitRequest:
      type: object
      required:
        - circuit
      properties:
        circuit:
          type: string
          minLength: 1
          description: Quantum circuit definition (QASM3 or other text format)
          example: "OPENQASM 3; qubit q; h q; measure q;"
      example:
        circuit: "OPENQASM 3; qubit q; h q; measure q;"

    TaskSubmitResponse:
      type: object
      required:
        - task_id
        - message
        - correlation_id
      properties:
        task_id:
          type: string
          format: uuid
          description: Unique identifier for the submitted task
          example: "550e8400-e29b-41d4-a716-446655440000"
        message:
          type: string
          description: Human-readable confirmation message
          example: "Task submitted successfully."
        correlation_id:
          type: string
          format: uuid
          description: Request correlation ID for tracing
          example: "abc123-def456-789012"

    TaskStatusResponse:
      type: object
      required:
        - status
        - correlation_id
      properties:
        status:
          type: string
          enum:
            - pending
            - processing
            - completed
            - failed
          description: Current task state
          example: "pending"
        message:
          type: string
          description: Human-readable status description (present for pending/failed states)
          example: "Task is still in progress."
        result:
          type: object
          description: Task execution results (present only for completed tasks)
          example:
            "0": 512
            "1": 512
        correlation_id:
          type: string
          format: uuid
          description: Request correlation ID for tracing
          example: "xyz789-uvw456"
      example:
        status: "pending"
        message: "Task is still in progress."
        correlation_id: "xyz789-uvw456"

    HealthCheckResponse:
      type: object
      required:
        - status
        - timestamp
      properties:
        status:
          type: string
          enum:
            - healthy
            - degraded
            - unavailable
          description: Service health status
          example: "healthy"
        timestamp:
          type: string
          format: date-time
          description: UTC timestamp of health check
          example: "2025-12-28T12:00:00Z"

    ErrorResponse:
      type: object
      required:
        - error
        - correlation_id
      properties:
        error:
          type: string
          description: Brief error description
          example: "Validation failed"
        details:
          type: object
          additionalProperties: true
          description: Field-specific validation errors (present for 400 errors)
          example:
            circuit: "Field required"
        correlation_id:
          type: string
          format: uuid
          description: Request correlation ID for log tracing
          example: "abc123-def456"

  headers:
    X-Correlation-ID:
      description: Correlation ID for request tracing across services
      schema:
        type: string
        format: uuid
      example: "abc123-def456-789012"

  securitySchemes:
    # Future: API Key or JWT authentication
    # For stub phase, no authentication required
    {}